references/alpaca-crypto-streaming/requirements.txt
# Crypto Streaming Requirements
# Phase 3 of Pulse Element Project

# Core WebSocket client
websocket-client>=1.3.0

# JSON handling (built-in to Python, listed for completeness)  
# json - built-in

# Date/time handling (built-in to Python)
# datetime - built-in

# Threading support (built-in to Python)
# threading - built-in

# Optional: For enhanced WebSocket performance
# websockets>=10.0  # Alternative WebSocket library (not required)

# For testing and development
pytest>=7.0.0